backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Document Processing
pypdf>=3.17.0
python-docx==1.1.0
pdfplumber==0.10.3

# AI/ML Libraries
spacy==3.7.2
transformers==4.36.2
torch>=2.2.0
scikit-learn==1.3.2
numpy==1.26.2

# NLP & Text Processing
nltk==3.8.1
textblob==0.17.1

# PDF Generation
reportlab==4.0.7
fpdf2==2.7.6

# Utilities
Werkzeug==3.0.1
python-magic==0.4.27
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
